requirements.txt
python-telegram-bot==20.7
matplotlib>=3.0.0
requests>=2.25.0
pytz>=2021.1
openai>=1.0.0
python-dotenv>=1.0.0